Cleaning the aligned TAD borders score from TADbit
The TAD reproducibility score is the number of times a TAD is called by the algorithm. As the algorithm of TAD calling is carried out 10 times, TAD reproducibility values range from 1 to 10. TADs with a score lower than 5 will be considered not robust and are likely to be undetectable in replicates or at different resolutions.